requirements.txt
# Trinity 3.0 Dependencies
# Production-ready deployment requirements

# Core Framework
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.18.0

# Intelligence Layer (Entity Extraction)
instructor>=1.0.0
pydantic>=2.0.0
anthropic>=0.25.0

# Data Integration
openbb>=4.0.0
requests>=2.31.0

# Database (for prediction tracking)
psycopg2-binary>=2.9.0

# Graph Backend
networkx>=3.1

# Caching (Optional)
redis>=5.0.0

# Utilities
python-dotenv>=1.0.0

# Testing (Development only)
pytest>=7.4.0
pytest-asyncio>=0.21.0
